pizzalist functional - working on pizzalistitem and add/remve button
