+++ src/components/App/App.jsx
@@ -12,9 +12,34 @@ import Header from '../Header/Header';
 import PizzaForm from '../PizzaForm/PizzaForm';
 import PizzaList from '../PizzaList/PizzaList';
 import TotalCost from '../TotalCost/TotalCost';
+import {useEffect} from 'react';
+import {useDispatch} from 'react-redux';
+
 
 
 function App() {
+
+      const dispatch = useDispatch();
+
+
+    useEffect(() => {
+      getOptions();
+    },[])
+
+  const getOptions = () => {
+    axios({
+        method: 'GET',
+        url: '/api/pizza'
+    }).then(response => { //response is array of food options
+        console.log(response.data);
+        dispatch({
+            type: 'SHOW_PIZZAS',
+            payload: response.data
+        })  
+    }).catch(error => {
+        console.log('error in pizzalistitem get', error);
+    })
+}
 
   return (
     <Router>
